backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
colorama==0.4.6
Django==6.0.1
django-cors-headers==4.9.0
django-leaflet==0.33.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
geopandas==1.1.2
gunicorn==23.0.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
psycopg2==2.9.11
PyJWT==2.10.1
pyogrio==0.12.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
